chore: relocate backup-wallet screen files
