+++ components/brave_wallet_ui/page/screens/backup-wallet/backup-recovery-phrase/onboarding-backup-recovery-phrase.stories.tsx
