+++ components/brave_wallet_ui/page/screens/backup-wallet/verify-recovery-phrase/verify-backup-recovery-phrase.style.ts
